Refactor class namespaces
