+++ classes/local/clients/system_folder_access.php
@@ -25,7 +25,7 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace mod_collaborativefolders;
+namespace mod_collaborativefolders\local\clients;
 
 defined('MOODLE_INTERNAL') || die();
